Refactor to permit base-on of flattened SquashFS only images.
Use unsquashfs to decompress flattened image directly to the
  _instroot.
Avoid kickstart installations when a change root shell is requested
  in a livecd-creator session that employs a base-on image.  This
  makes sequential editing of an image quicker.

Also:
Include F2FS drivers in the initrd.
Default to zstd compression to favor more rapid decompression.
Include the dracut modules dmsquash-live-autooverlay & overlayfs
  among the defaults.
Revert the explicit omission of plymouth.
Rename dracut_args to dracut_conf_args to match their required
  syntax.
Print a couple of console messages on the activity of the script at
  longer operations.
Assure that mounts are unmounted in the necessary order.
Update documentation.

+++ imgcreate/creator.py
@@ -481,6 +481,7 @@ class ImageCreator(object):
         os.umask(origumask)
 
     def __load_selinuxfs(self):
+        print('Setting SELinux contexts...')
         arglist = ["mount", "--bind", "/dev/null",
                    self._instroot + self.__selinux_mountpoint + "/load"]
         subprocess.call(arglist, close_fds = True)
@@ -570,7 +571,8 @@ class ImageCreator(object):
 
         self.__create_minimal_dev()
 
-        os.symlink("/proc/self/mounts", self._instroot + "/etc/mtab")
+        if not os.path.lexists(self._instroot + "/etc/mtab"):
+            os.symlink("/proc/self/mounts", self._instroot + "/etc/mtab")
 
         self.__write_fstab()
 
@@ -1001,15 +1003,15 @@ class LoopImageCreator(ImageCreator):
     def _mount_instroot(self, base_on = None):
         self.__imgdir = self._mkdtemp()
 
-        if not base_on is None:
-            self._base_on(base_on)
-
         self.__instloop = ExtDiskMount(SparseLoopbackDisk(self._image,
                                                           self.__image_size),
                                        self._instroot,
                                        self._fstype,
                                        self.__blocksize,
                                        self.fslabel)
+
+        if not base_on is None:
+            self._base_on(base_on)
 
         try:
             self.__instloop.mount()
